03-langchain-latest/requirements.txt
streamlit==1.31.0
langchain-core==0.1.28
langchain-openai==0.0.5
langchain-community==0.0.21
langchain==0.1.9
langchain-experimental==0.0.49
faiss-cpu==1.7.4
sqlite3-api==0.1.0
pandas==2.1.4
matplotlib==3.8.2
yfinance==0.2.37
python-dotenv==1.0.1
tavily-python==0.3.1
pydantic==2.6.3
openai==1.12.0 